+++ include/libpmemobj++/allocator.hpp
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016-2018, Intel Corporation
+ * Copyright 2016-2019, Intel Corporation
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
@@ -40,8 +40,8 @@
 
 #include <libpmemobj++/detail/common.hpp>
 #include <libpmemobj++/detail/life.hpp>
-#include <libpmemobj++/detail/pexceptions.hpp>
 #include <libpmemobj++/persistent_ptr.hpp>
+#include <libpmemobj++/pexceptions.hpp>
 #include <libpmemobj++/pext.hpp>
 #include <libpmemobj.h>
