+++ include/libpmemobj++/mutex.hpp
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016-2018, Intel Corporation
+ * Copyright 2016-2019, Intel Corporation
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
@@ -38,7 +38,7 @@
 #ifndef LIBPMEMOBJ_CPP_MUTEX_HPP
 #define LIBPMEMOBJ_CPP_MUTEX_HPP
 
-#include <libpmemobj++/detail/pexceptions.hpp>
+#include <libpmemobj++/pexceptions.hpp>
 #include <libpmemobj/thread.h>
 #include <libpmemobj/tx_base.h>
