+++ include/libpmemobj++/pool.hpp
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016-2018, Intel Corporation
+ * Copyright 2016-2019, Intel Corporation
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
@@ -44,8 +44,8 @@
 
 #include <libpmemobj++/detail/common.hpp>
 #include <libpmemobj++/detail/ctl.hpp>
-#include <libpmemobj++/detail/pexceptions.hpp>
 #include <libpmemobj++/p.hpp>
+#include <libpmemobj++/pexceptions.hpp>
 #include <libpmemobj/pool_base.h>
 
 namespace pmem
