+++ include/libpmemobj++/utils.hpp
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016-2018, Intel Corporation
+ * Copyright 2016-2019, Intel Corporation
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
@@ -37,8 +37,8 @@
 #ifndef LIBPMEMOBJ_CPP_UTILS_HPP
 #define LIBPMEMOBJ_CPP_UTILS_HPP
 
-#include <libpmemobj++/detail/pexceptions.hpp>
 #include <libpmemobj++/persistent_ptr.hpp>
+#include <libpmemobj++/pexceptions.hpp>
 #include <libpmemobj/base.h>
 
 namespace pmem
